Add files via upload
In this project, I implemented a ML model that will try to recognize the handwritten text extracted from the popular MNIST dataset with a high degree of accuracy. For this purpose, I’ve created and trained a Neural Network model that, after the training, will be able to predict digits from hand-written images with a high degree of accuracy.

Solution Approach:
The following solution approach for the implementation of the HTR (handwritten text recognition) model is followed:
1. Import MNIST dataset
2. One hot encoding
I changed the way label is represented from a class name or number to a list of all possible classes with all the classes set to 0 except the one which this example belongs to - which will be set to 1.
3. Preprocessing the examples
I created a Neural Network which will take 784-dimensional vectors as inputs (28 rows * 28 columns) and will output a 10-dimensional vector (For the 10 classes). I have already converted the outputs to 10 dimensional, one-hot encoded vectors. I used numpy to easily unroll the examples from (28, 28) arrays to (784, 1) vectors.
4. Creating a model
I used a Sequential class defined in Keras to create our model. All the layers are going to be Dense layers. This means, like our examples, all the nodes of a layer would be connected to all the nodes of the preceding layer i.e. densely connected.
5. Training the model
I trained the model for 3 epochs. Think of epoch like an iteration of all the examples going through the model. So, by setting the epochs to 3, we will go through all the training examples 3 times.
6. Prediction
In this, I plotted the first few test set images along with their predicted and actual labels and see how our trained model actually performed.